pyproject.toml
[project]
name = "auto-trading-bot"
version = "0.1.0"
description = "Claude Opus-powered automated trading bot for G7 exchange-traded instruments"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "yfinance>=0.2.40",
    "fredapi>=0.5.2",
    "finnhub-python>=2.4.20",
    "vaderSentiment>=3.3.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
